IMPORTANT!!! some changes of player and hurt box code
IMPORTANT!!!

I changed the name of the player's hurt box Collision since var collistion store the node of hurt box collision, if the name of hurt box collision player and mob are not the same, it must have error when running the game since one of them will not get the Hurt Box Collision.

if we want to let the player and mob use the same hurtbox.gd, the name of the hurt box collision of both of them must be the same.

I also add some comments in the hurt box and hitbox code, see them.